refactor: move constraints
