+++ packages/components/constraints/src/horizontal/horizontal.visualroute.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from '@emotion/styled';
 import { Constraints } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const GreenBox = styled.div`
   width: 100%;
