+++ src/components/label/label.story.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, boolean, text, select } from '@storybook/addon-knobs/react';
-import Constraints from '../constraints';
+import Constraints from '@commercetools-uikit/constraints';
 import Section from '../../../.storybook/decorators/section';
 import Readme from './README.md';
 import Label from './label';
